***PLEASE READ***
The "Documents" page of stevenson.space does not work, google returns a Error 401: deleted_client error. After some investigation, this is mostly likely because the project was deleted inside google developer console. Since issues are disabled in this repo, I have made a pull request with a new working auth key. I would recommend creating a new oath 2.0 key inside the console and replaceing it with the one I have in this project.