**important** schedule fix for today
I found a bug relating the schedule and cookies. The schuedle showed was the one in cookies, not the updated one in the json. the thanksgiving schedule was off because It had the same name (I think).  This is for today, if you can't get to this, that is ok. I had no idea this would happen. Thanks - Joey