Create SHS schedule display for SHS bulletin boards (TVSpace) (#283)
* Create very basic page for display on TV

* Don't display ScheduleCard title if given title is empty string

To make ScheduleCard usable for tvspace. I'm intending to display the countdown timer there instead, so having the title not render at all if desired is helpful.

* Move internals of ScheduleCard into ScrollablePeriodList.vue

Mainly, this is for more flexibility of adding things to the card in ScheduleCard. Currently, the slot is at the bottom, but I'm intending to add other things at the top for tvspace, but I don't want it to interfere with the code that expects the slot to be at the bottom.

* Replace ScheduleCard in TVSpace display with scrollable period list

* Add styling to countdown and schedule name

* Scroll to current period when height changes

* Fix scrolling on tvspace

FIXME: the max-height is a hacky fix until the cards & period component can be fixed to be more responsive. currently, the scrolling for the period list doesn't display unless the container is constrained smaller than the max size it can be, and 95vh is mostly unnoticeable but gets the scrolling working

* Display divider between countdown and period list

* Increase font size slightly

* Increase font size for period list

* Embolden active range using not selector

* Attempt to make text more readable

* Allow "never" as a theme schedule

TVSpace theme should not be displayed to users

* Create TVSpace theme

* Brighten schedule name

* Automatically set TVSpace theme if 'settheme' is set to true in route

* Remove unused import in ScrollablePeriodList.vue